Extract useUserSettledBid
